backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
openai
pydantic
sse-starlette
selenium
apify-client
langchain
langchain-community
langchain_openai
beautifulsoup4